requirements.txt
# Core API Framework
fastapi==0.128.0
uvicorn==0.39.0

# Environment & Configuration
python-dotenv==1.2.1

# Database
supabase==2.27.2

# Authentication
PyJWT==2.10.1

# Data Validation
pydantic==2.12.5

# HTTP Client & Web Scraping (Custom Crawler)
httpx==0.28.1
beautifulsoup4==4.14.3
lxml==6.0.2
brotli==1.2.0  # Required for Brotli compression support

# Browser Automation (Cloudflare Bypass) - DISABLED FOR DEPLOYMENT
# playwright==1.58.0

# Scalability - Distributed Task Queue & State Management
celery==5.4.0
redis==5.2.1
flower==2.0.1  # Celery monitoring dashboard

# AI - Google Gemini API for redirect suggestions
google-generativeai==0.8.3